pyproject.toml
[tool.poetry]
name = "fogtime"
version = "0.1.0"
description = "A small script that syncs appointments from multipe google calenders as blockers to a single one."
authors = ["Robert Franzke <r.l.franzke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.169.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
